gpg.py: Add listkey() and listkeys() functions
Add functions for handling --list-key and --list-keys. Both of these functions
have the same signature as the fingerprint() method. listkey() and listkeys()
both supply arguments after --list-key and --list-keys as inputfile parameter
to runGPG(). The inputfile parameter may be the name of a filename parameter
to gpg or a non-filename arguments supplied to --list-keys or --list-key such
as a user's email address (uid), key id (pub/sub) or key fingerprint.